# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RescuEat API - Food rescue platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  store_owner: "store_owner",
  admin: "admin",
} as const;

export interface User {
  id: string;
  name: string;
  email?: string;
  role: UserRole;
  createdAt?: string;
}

export type StoreCategory = (typeof StoreCategory)[keyof typeof StoreCategory];

export const StoreCategory = {
  restaurant: "restaurant",
  bakery: "bakery",
  cafe: "cafe",
  supermarket: "supermarket",
  convenience: "convenience",
  other: "other",
} as const;

export interface Store {
  id: number;
  name: string;
  description?: string;
  address: string;
  city: string;
  category: StoreCategory;
  lat: number;
  lng: number;
  imageUrl?: string;
  phone?: string;
  openTime?: string;
  closeTime?: string;
  rating?: number;
  isActive: boolean;
  ownerId?: string;
  totalBagsAvailable: number;
  createdAt: string;
}

export type CreateStoreRequestCategory =
  (typeof CreateStoreRequestCategory)[keyof typeof CreateStoreRequestCategory];

export const CreateStoreRequestCategory = {
  restaurant: "restaurant",
  bakery: "bakery",
  cafe: "cafe",
  supermarket: "supermarket",
  convenience: "convenience",
  other: "other",
} as const;

export interface CreateStoreRequest {
  name: string;
  description?: string;
  address: string;
  city: string;
  category: CreateStoreRequestCategory;
  lat: number;
  lng: number;
  imageUrl?: string;
  phone?: string;
  openTime?: string;
  closeTime?: string;
}

export type UpdateStoreRequestCategory =
  (typeof UpdateStoreRequestCategory)[keyof typeof UpdateStoreRequestCategory];

export const UpdateStoreRequestCategory = {
  restaurant: "restaurant",
  bakery: "bakery",
  cafe: "cafe",
  supermarket: "supermarket",
  convenience: "convenience",
  other: "other",
} as const;

export interface UpdateStoreRequest {
  name?: string;
  description?: string;
  address?: string;
  city?: string;
  category?: UpdateStoreRequestCategory;
  imageUrl?: string;
  phone?: string;
  openTime?: string;
  closeTime?: string;
  isActive?: boolean;
}

export interface SurpriseBag {
  id: number;
  storeId: number;
  title: string;
  description?: string;
  originalPrice: number;
  discountedPrice: number;
  stockCount: number;
  pickupStart?: string;
  pickupEnd?: string;
  isActive: boolean;
  createdAt: string;
}

export type SurpriseBagWithStore = SurpriseBag & {
  store: Store;
};

export interface CreateBagRequest {
  title: string;
  description?: string;
  originalPrice: number;
  discountedPrice: number;
  stockCount: number;
  pickupStart?: string;
  pickupEnd?: string;
}

export interface UpdateBagRequest {
  title?: string;
  description?: string;
  originalPrice?: number;
  discountedPrice?: number;
  stockCount?: number;
  pickupStart?: string;
  pickupEnd?: string;
  isActive?: boolean;
}

export type ReservationStatus =
  (typeof ReservationStatus)[keyof typeof ReservationStatus];

export const ReservationStatus = {
  pending: "pending",
  confirmed: "confirmed",
  picked_up: "picked_up",
  cancelled: "cancelled",
} as const;

export type ReservationPaymentStatus =
  (typeof ReservationPaymentStatus)[keyof typeof ReservationPaymentStatus];

export const ReservationPaymentStatus = {
  unpaid: "unpaid",
  paid: "paid",
  refunded: "refunded",
} as const;

export interface Reservation {
  id: number;
  userId: string;
  bagId: number;
  storeId: number;
  quantity: number;
  totalPrice: number;
  status: ReservationStatus;
  paymentIntentId?: string;
  paymentStatus: ReservationPaymentStatus;
  pickupCode?: string;
  bag?: SurpriseBag;
  store?: Store;
  createdAt: string;
}

export interface CreateReservationRequest {
  bagId: number;
  quantity: number;
  userId: string;
}

export type UpdateReservationStatusRequestStatus =
  (typeof UpdateReservationStatusRequestStatus)[keyof typeof UpdateReservationStatusRequestStatus];

export const UpdateReservationStatusRequestStatus = {
  pending: "pending",
  confirmed: "confirmed",
  picked_up: "picked_up",
  cancelled: "cancelled",
} as const;

export interface UpdateReservationStatusRequest {
  status: UpdateReservationStatusRequestStatus;
}

export interface CreatePaymentIntentRequest {
  reservationId: number;
  bagId?: number;
  quantity?: number;
}

export interface PaymentIntentResponse {
  clientSecret: string;
  paymentIntentId: string;
  amount: number;
  currency: string;
}

export interface ConfirmPaymentRequest {
  reservationId: number;
  paymentIntentId: string;
}

export type ListStoresParams = {
  lat?: number;
  lng?: number;
  /**
   * Radius in km
   */
  radius?: number;
};

export type ListReservationsParams = {
  userId?: string;
  storeId?: number;
  status?: ListReservationsStatus;
};

export type ListReservationsStatus =
  (typeof ListReservationsStatus)[keyof typeof ListReservationsStatus];

export const ListReservationsStatus = {
  pending: "pending",
  confirmed: "confirmed",
  picked_up: "picked_up",
  cancelled: "cancelled",
} as const;