# lib/api-zod/src/generated/types/confirmPaymentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RescuEat API - Food rescue platform
 * OpenAPI spec version: 0.1.0
 */

export interface ConfirmPaymentRequest {
  reservationId: number;
  paymentIntentId: string;
}