# lib/api-zod/src/generated/types/createBagRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RescuEat API - Food rescue platform
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBagRequest {
  title: string;
  description?: string;
  originalPrice: number;
  discountedPrice: number;
  stockCount: number;
  pickupStart?: string;
  pickupEnd?: string;
}