# lib/api-zod/src/generated/types/createReservationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RescuEat API - Food rescue platform
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReservationRequest {
  bagId: number;
  quantity: number;
  userId: string;
}